Add files via upload
Inspiration- The inspiration for the project came from the significant child mortality rate of the world and specially in African and Asian regions where the child mortality rate is as high as 114 deaths per 1000 births.The project will help to visualise and predict the child mortality rate and take major steps to decrease the child mortality rate.
What it does-The project helps to predict the child mortality rate of the world and visualize the data using machine learning.It will help people to get transparent information of the data with the help of an web application and it will also be helpful in the field of healthcare and pediatric care.It will also serve as a reference to the researcher for future research.
How we built it-The project was built by using the concept of python and machine learning.
Challenges we ran into-One of the major challenges were collecting the suitable dataset for this project from a trustworthy source, implementation of our ideas correctly and choosing the right algorithm for the model for achieving the highest accuracy possible for the machine learning model.We have to test the model multiple times using multiple algorithms before finalizing the final model.
Accomplishments that we're proud of- We are proud the we were able to achieve our initial goal and build a model that can help the healthcare system and the society towards its betterment.
What we learned-We have learned many new concepts and ways in which we can improve our model, make some more models with better use of machine learning and how we can implement our idea in real world for the betterment of society.
What's next for World Child Mortality Rate Data Visualization And Prediction-We are planning to update the model to take real time data and predict the mortality rate based on that and simultaneously giving a vivid and precise Visualization of the data. We are also planning to develop the frontend more and display the major causes of such high mortality rate and how can we prevent them and ways to reduce them.